Split runner file read tests
